+++ CHANGELOG.md
@@ -10,6 +10,7 @@ project adheres to
 ## Unreleased
 
 * Improve support for the `selector.nest` function (PR #189).
+* Move `selector.append` to new selector implementation (PR #190).
 * Some internal cleanup and improvements in the next-generation css
   selector implementation (which is currently internal and used only
   for selector functions, but should replace the old css selector